[Refactor] Refactor ANF partition pass to extract some common utilities.
